requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
python-dotenv==1.2.2
sentence-transformers==5.3.0
pgvector==0.4.2
python-multipart==0.0.22
streamlit==1.56.0
requests==2.33.1